+++ src/ai/backend/gateway/admin.py
@@ -1,4 +1,3 @@
-import asyncio
 import inspect
 import logging
 from typing import (
@@ -23,13 +22,28 @@ from .manager import GQLMutationUnfrozenRequiredMiddleware
 from .exceptions import GraphQLError as BackendGQLError
 from .auth import auth_required
 from .typing import CORSOptions, WebMiddleware
-from .utils import check_api_params, trim_text
+from .utils import check_api_params
 from ..manager.models.base import DataLoaderManager
-from ..manager.models.gql import Mutations, Queries
+from ..manager.models.gql import (
+    Mutations, Queries,
+    GQLMutationPrivilegeCheckMiddleware,
+)
 
 log = BraceStyleAdapter(logging.getLogger('ai.backend.gateway.admin'))
 
 _rx_qname = re.compile(r'{\s*(\w+)\b')
+
+
+class GQLLoggingMiddleware:
+
+    def resolve(self, next, root, info, **args):
+        if len(info.path) == 1:
+            log.info('ADMIN.GQL (ak:{}, {}:{}, op:{})',
+                     info.context['access_key'],
+                     info.operation.operation,
+                     info.field_name,
+                     info.operation.name)
+        return next(root, info, **args)
 
 
 @atomic
@@ -45,14 +59,6 @@ async def handle_gql(request: web.Request, params: Any) -> web.Response:
     schema = request.app['admin.gql_schema']
     manager_status = await request.app['config_server'].get_manager_status()
     known_slot_types = await request.app['config_server'].get_resource_slots()
-    match = _rx_qname.search(params['query'].replace('\n', ''))
-    if match:
-        qsummary = match.group(1)
-    else:
-        qsummary = trim_text(params['query'], 80)
-    log.info('ADMIN.GQL (ak:{}, query:{!r}, op:{})',
-             request['keypair']['access_key'],
-             qsummary, params['operation_name'])
     context = {
         'config': request.app['config'],
         'config_server': request.app['config_server'],
@@ -73,7 +79,11 @@ async def handle_gql(request: web.Request, params: Any) -> web.Response:
             'dlmgr': dlmanager,
             **context,
         },
-        middleware=[GQLMutationUnfrozenRequiredMiddleware()],
+        middleware=[
+            GQLLoggingMiddleware(),
+            GQLMutationUnfrozenRequiredMiddleware(),
+            GQLMutationPrivilegeCheckMiddleware(),
+        ],
         return_promise=True)
     if inspect.isawaitable(result):
         result = await result
@@ -89,8 +99,7 @@ async def handle_gql(request: web.Request, params: Any) -> web.Response:
 
 
 async def init(app: web.Application) -> None:
-    loop = asyncio.get_event_loop()
-    app['admin.gql_executor'] = AsyncioExecutor(loop=loop)
+    app['admin.gql_executor'] = AsyncioExecutor()
     app['admin.gql_schema'] = graphene.Schema(
         query=Queries,
         mutation=Mutations,