+++ src/ai/backend/gateway/server.py
@@ -32,12 +32,14 @@ from setproctitle import setproctitle
 
 from ai.backend.common import redis
 from ai.backend.common.cli import LazyGroup
-from ai.backend.common.utils import env_info
+from ai.backend.common.config import redis_config_iv
+from ai.backend.common.utils import env_info, current_loop
 from ai.backend.common.logging import Logger, BraceStyleAdapter
 from ai.backend.common.plugin import (
     discover_entrypoints, install_plugins
 )
 from ..manager import __version__
+from ..manager.plugin.error_monitor import ErrorMonitor
 from ..manager.registry import AgentRegistry
 from ..manager.scheduler.dispatcher import SchedulerDispatcher
 from ..manager.background import BackgroundTaskManager
@@ -167,7 +169,7 @@ async def exception_middleware(request: web.Request,
             log.exception('Internal server error raised inside handlers')
             raise
         if (error_monitor := app.get('error_monitor', None)) is not None:
-            error_monitor.capture_exception()
+            await error_monitor.capture_exception()
         if (stats_monitor := app.get('stats_monitor', None)) is not None:
             stats_monitor.report_stats('increment', 'ai.backend.gateway.api.failures')
             stats_monitor.report_stats('increment', f'ai.backend.gateway.api.status.{ex.status_code}')
@@ -190,7 +192,7 @@ async def exception_middleware(request: web.Request,
         raise e
     except Exception as e:
         if (error_monitor := app.get('error_monitor', None)) is not None:
-            error_monitor.capture_exception()
+            await error_monitor.capture_exception()
         log.exception('Uncaught exception in HTTP request handlers {0!r}', e)
         if app['config']['debug']['enabled']:
             raise InternalServerError(traceback.format_exc())
@@ -209,10 +211,12 @@ async def config_server_ctx(app: web.Application) -> AsyncIterator[None]:
         app['config']['etcd']['user'], app['config']['etcd']['password'],
         app['config']['etcd']['namespace'])
 
-    shared_config = await load_shared_config(app['config_server'].etcd)
-    app['config'].update(shared_config)
-    redis_config = await app['config_server'].etcd.get_prefix('config/redis')
-    app['config']['redis'] = redis_config_iv.check(redis_config)
+    app['config'].update(
+        await load_shared_config(app['config_server'].etcd)
+    )
+    app['config']['redis'] = redis_config_iv.check(
+        await app['config_server'].etcd.get_prefix('config/redis')
+    )
     _update_public_interface_objs(app)
     yield
     # await app['config_server'].close()
@@ -308,9 +312,9 @@ async def sched_dispatcher_ctx(app: web.Application) -> AsyncIterator[None]:
 
 async def monitoring_ctx(app: web.Application) -> AsyncIterator[None]:
     # Install stats hook plugins.
+    app['error_monitor'] = ErrorMonitor(app)
     plugins = [
         'stats_monitor',
-        'error_monitor',
     ]
     install_plugins(plugins, app, 'dict', app['config'])
     _update_public_interface_objs(app)
@@ -361,24 +365,24 @@ async def webapp_plugins_ctx(app: web.Application) -> AsyncIterator[None]:
 
 
 def handle_loop_error(
-    app: web.Application,
+    root_app: web.Application,
     loop: asyncio.AbstractEventLoop,
     context: Mapping[str, Any],
 ) -> None:
+    if isinstance(loop, aiojobs.Scheduler):
+        loop = current_loop()
     exception = context.get('exception')
     msg = context.get('message', '(empty message)')
     if exception is not None:
-        if (error_monitor := app.get('error_monitor', None)) is not None:
-            error_monitor.set_context(context)
         if sys.exc_info()[0] is not None:
             log.exception('Error inside event loop: {0}', msg)
-            if error_monitor is not None:
-                error_monitor.capture_exception(True)
+            if error_monitor := root_app.get('error_monitor', None):
+                loop.create_task(error_monitor.capture_exception())
         else:
             exc_info = (type(exception), exception, exception.__traceback__)
             log.error('Error inside event loop: {0}', msg, exc_info=exc_info)
-            if error_monitor is not None:
-                error_monitor.capture_exception(exc_info)
+            if error_monitor := root_app.get('error_monitor', None):
+                loop.create_task(error_monitor.capture_exception(exception))
 
 
 def _init_subapp(pkg_name: str,
@@ -447,11 +451,11 @@ def build_root_app(pidx: int,
     app.on_response_prepare.append(on_prepare)
     if cleanup_contexts is None:
         app.cleanup_ctx.append(config_server_ctx)
-        app.cleanup_ctx.append(monitoring_ctx)
         app.cleanup_ctx.append(manager_status_ctx)
         app.cleanup_ctx.append(redis_ctx)
         app.cleanup_ctx.append(database_ctx)
         app.cleanup_ctx.append(event_dispatcher_ctx)
+        app.cleanup_ctx.append(monitoring_ctx)
         app.cleanup_ctx.append(agent_registry_ctx)
         app.cleanup_ctx.append(sched_dispatcher_ctx)
         app.cleanup_ctx.append(webapp_plugins_ctx)
@@ -490,6 +494,7 @@ async def server_main(loop: asyncio.AbstractEventLoop,
         '.session_template',
         '.image',
         '.userconfig',
+        '.logs',
     ]
     app = build_root_app(pidx, _args[0], subapp_pkgs=subapp_pkgs)