+++ src/ai/backend/gateway/userconfig.py
@@ -19,9 +19,14 @@ from .typing import CORSOptions, Iterable, WebMiddleware
 from .utils import check_api_params, get_access_key_scopes
 
 from ..manager.models import (
-    keypairs, verify_dotfile_name
+    keypairs,
+    vfolders,
+    query_accessible_vfolders,
+    query_bootstrap_script,
+    query_owned_dotfiles,
+    verify_dotfile_name,
+    MAXIMUM_DOTFILE_SIZE,
 )
-from ..manager.models.keypair import query_owned_dotfiles, MAXIMUM_DOTFILE_SIZE
 
 log = BraceStyleAdapter(logging.getLogger('ai.backend.gateway.dotfile'))
 
@@ -41,7 +46,7 @@ async def create(request: web.Request, params: Any) -> web.Response:
     log.info('CREATE (ak:{0}/{1})',
              requester_access_key, owner_access_key if owner_access_key != requester_access_key else '*')
     dbpool = request.app['dbpool']
-
+    user_uuid = request['user']['uuid']
     async with dbpool.acquire() as conn, conn.begin():
         path: str = params['path']
         dotfiles, leftover_space = await query_owned_dotfiles(conn, owner_access_key)
@@ -50,7 +55,11 @@ async def create(request: web.Request, params: Any) -> web.Response:
         if len(dotfiles) == 100:
             raise DotfileCreationFailed('Dotfile creation limit reached')
         if not verify_dotfile_name(path):
-            raise InvalidAPIParameters(f'Dotfile {path} is reserved for internal operations.')
+            raise InvalidAPIParameters('dotfile path is reserved for internal operations.')
+        duplicate_vfolder = \
+            await query_accessible_vfolders(conn, user_uuid, extra_vf_conds=(vfolders.c.name == path))
+        if len(duplicate_vfolder) > 0:
+            raise InvalidAPIParameters('dotfile path conflicts with your dot-prefixed vFolder')
         duplicate = [x for x in dotfiles if x['path'] == path]
         if len(duplicate) > 0:
             raise DotfileAlreadyExists
@@ -163,6 +172,39 @@ async def delete(request: web.Request, params: Any) -> web.Response:
         return web.json_response({'success': True})
 
 
+@server_status_required(READ_ALLOWED)
+@auth_required
+@check_api_params(t.Dict(
+    {
+        t.Key('script'): t.String(max_length=MAXIMUM_DOTFILE_SIZE),
+    }
+))
+async def update_bootstrap_script(request: web.Request, params: Any) -> web.Response:
+    dbpool = request.app['dbpool']
+    access_key = request['keypair']['access_key']
+    log.info('UPDATE_BOOTSTRAP_SCRIPT (ak:{0})', access_key)
+    async with dbpool.acquire() as conn, conn.begin():
+        script = params.get('script', '').strip()
+        if len(script > MAXIMUM_DOTFILE_SIZE):
+            raise DotfileCreationFailed('Maximum bootstrap script length reached')
+        query = (keypairs.update()
+                         .values(bootstrap_script=script)
+                         .where(keypairs.c.access_key == access_key))
+        await conn.execute(query)
+    return web.json_response({})
+
+
+@auth_required
+@server_status_required(READ_ALLOWED)
+async def get_bootstrap_script(request: web.Request) -> web.Response:
+    dbpool = request.app['dbpool']
+    access_key = request['keypair']['access_key']
+    log.info('GET_BOOTSTRAP_SCRIPT (ak:{0})', access_key)
+    async with dbpool.acquire() as conn:
+        script, _ = await query_bootstrap_script(conn, access_key)
+        return web.json_response(script)
+
+
 async def init(app: web.Application) -> None:
     pass
 
@@ -182,5 +224,7 @@ def create_app(default_cors_options: CORSOptions) -> Tuple[web.Application, Iter
     cors.add(app.router.add_route('GET', '/dotfiles', list_or_get))
     cors.add(app.router.add_route('PATCH', '/dotfiles', update))
     cors.add(app.router.add_route('DELETE', '/dotfiles', delete))
+    cors.add(app.router.add_route('POST', '/bootstrap-script', update_bootstrap_script))
+    cors.add(app.router.add_route('GET', '/bootstrap-script', get_bootstrap_script))
 
     return app, []